tests/wpt/css-tests/cssom-view-1_dev/html/CaretPosition-001.htm

<!DOCTYPE html>
<html><head>
	<title>CSSOM View Module test:CaretPosition</title>
    <link href="mailto:tidelgl@gmail.com" rel="author" title="unbug">
    <link href="http://www.w3.org/TR/cssom-view/#the-caretposition-interface" rel="help">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>	
	<style type="text/css">
	.box{
	    position: absolute;
	    top : 100px;
	    left : 400px;
	    width: 100px;
	    height: 100px;
	    background-color: #000;
	}
	</style>
</head>
<body>
	<p>This case tests the Screen pixelDepth and colorDepth</p>
	<p>The test passes if the value is Element "#box2"</p>
	<div class="box" id="box"></div>
	<div class="box"></div>
	<div class="box"></div>
	<div class="box"></div>
	<div class="box"></div>
	<div class="box"></div>
	<div class="box" id="box2"></div>
	<div style="left: 405px;background:red;" class="box" id="box3"></div>
	<div id="log"></div>
	<script>
        test(function(){
            assert_equals(getBox('box2'), document.elementFromPoint(400,100), "Expected value for element id is 'box2'");
        },'getBox');
        test(function(){
            assert_equals(null, document.elementFromPoint(400,900), "Expected value for element id is 'box2'");
        },'getBox');
        function getBox(id){
            return document.getElementById(id);
        }	    
	</script>

</body></html>
